Moved to a folder outside test.
