+++ src/QirRuntime/CMakeLists.txt
@@ -121,5 +121,5 @@ endif()
 
 
 add_subdirectory(lib)
+add_subdirectory(samples)
 add_subdirectory(test)
-
