+++ src/QirRuntime/samples/CMakeLists.txt
@@ -1,0 +1,3 @@
+set(standalone_includes ""${PROJECT_SOURCE_DIR}/externals/CLI11"")
+
+add_subdirectory(StandaloneSupportedInputs)
