+++ src/QirRuntime/samples/StandaloneSupportedInputs/.clang-tidy
