+++ src/QirRuntime/samples/StandaloneSupportedInputs/qsharp/qir-entrypoint-supported-inputs.csproj
@@ -1,0 +1,9 @@
+<Project Sdk="Microsoft.Quantum.Sdk/0.15.2102128704-alpha">
+
+  <PropertyGroup>
+    <OutputType>Exe</OutputType>
+    <TargetFramework>netcoreapp3.1</TargetFramework>
+    <QirGeneration>True</QirGeneration> 
+  </PropertyGroup>
+
+</Project>
