--- src/QirRuntime/test/Standalone/CMakeLists.txt
@@ -1,2 +1,0 @@
-add_subdirectory(RandomBit)
-add_subdirectory(InputTypes)
