--- src/QirRuntime/test/Standalone/InputTypes/input-types.qs
@@ -1,11 +1,0 @@
-﻿// Copyright (c) Microsoft Corporation. All rights reserved.
-// Licensed under the MIT License.
-
-namespace Microsoft.Quantum.Testing.QIR {
-    open Microsoft.Quantum.Intrinsic;
-
-    @EntryPoint()
-    operation InputTypes(anInt : Int, aDouble : Double, anArray : Int[]) : Unit {
-        // TODO: Add messages with the values of the arguments.
-    }
-}
