--- src/QirRuntime/test/Standalone/RandomBit/.clang-tidy
@@ -1,3 +1,0 @@
-Checks: 
-  '-*,bugprone-*'
-WarningsAsErrors: '*'
