--- src/QirRuntime/test/Standalone/RandomBit/qir-random-bit.ll
@@ -1,108 +1,0 @@
-
-%Result = type opaque
-%Range = type { i64, i64, i64 }
-%Qubit = type opaque
-%Array = type opaque
-
-@ResultZero = external global %Result*
-@ResultOne = external global %Result*
-@PauliI = constant i2 0
-@PauliX = constant i2 1
-@PauliY = constant i2 -1
-@PauliZ = constant i2 -2
-@EmptyRange = internal constant %Range { i64 0, i64 1, i64 -1 }
-
-@Microsoft__Quantum__Testing__QIR__RandomBit = alias i1 (), i1 ()* @Microsoft__Quantum__Testing__QIR__RandomBit__body
-
-define i1 @Microsoft__Quantum__Testing__QIR__RandomBit__body() #0 {
-entry:
-  %q = call %Qubit* @__quantum__rt__qubit_allocate()
-  call void @__quantum__qis__h__body(%Qubit* %q)
-  %r = call %Result* @Microsoft__Quantum__Intrinsic__M__body(%Qubit* %q)
-  call void @__quantum__rt__qubit_release(%Qubit* %q)
-  call void @__quantum__rt__result_update_reference_count(%Result* %r, i64 -1)
-  ret i1 true
-}
-
-declare %Qubit* @__quantum__rt__qubit_allocate()
-
-declare %Array* @__quantum__rt__qubit_allocate_array(i64)
-
-declare void @__quantum__qis__h__body(%Qubit*)
-
-define %Result* @Microsoft__Quantum__Intrinsic__M__body(%Qubit* %qb) {
-entry:
-  %bases = call %Array* @__quantum__rt__array_create_1d(i32 1, i64 1)
-  %0 = call i8* @__quantum__rt__array_get_element_ptr_1d(%Array* %bases, i64 0)
-  %1 = bitcast i8* %0 to i2*
-  %2 = load i2, i2* @PauliZ
-  store i2 %2, i2* %1
-  call void @__quantum__rt__array_update_alias_count(%Array* %bases, i64 1)
-  %qubits = call %Array* @__quantum__rt__array_create_1d(i32 8, i64 1)
-  %3 = call i8* @__quantum__rt__array_get_element_ptr_1d(%Array* %qubits, i64 0)
-  %4 = bitcast i8* %3 to %Qubit**
-  store %Qubit* %qb, %Qubit** %4
-  call void @__quantum__rt__array_update_alias_count(%Array* %qubits, i64 1)
-  %5 = call %Result* @__quantum__qis__measure__body(%Array* %bases, %Array* %qubits)
-  call void @__quantum__rt__array_update_alias_count(%Array* %bases, i64 -1)
-  call void @__quantum__rt__array_update_alias_count(%Array* %qubits, i64 -1)
-  call void @__quantum__rt__array_update_reference_count(%Array* %bases, i64 -1)
-  call void @__quantum__rt__array_update_reference_count(%Array* %qubits, i64 -1)
-  ret %Result* %5
-}
-
-declare void @__quantum__rt__qubit_release(%Qubit*)
-
-declare void @__quantum__rt__result_update_reference_count(%Result*, i64)
-
-declare %Array* @__quantum__rt__array_create_1d(i32, i64)
-
-declare i8* @__quantum__rt__array_get_element_ptr_1d(%Array*, i64)
-
-declare void @__quantum__rt__array_update_alias_count(%Array*, i64)
-
-declare %Result* @__quantum__qis__measure__body(%Array*, %Array*)
-
-declare void @__quantum__rt__array_update_reference_count(%Array*, i64)
-
-define void @Microsoft__Quantum__Intrinsic__H__body(%Qubit* %qb) {
-entry:
-  call void @__quantum__qis__h__body(%Qubit* %qb)
-  ret void
-}
-
-define void @Microsoft__Quantum__Intrinsic__H__adj(%Qubit* %qb) {
-entry:
-  call void @__quantum__qis__h__body(%Qubit* %qb)
-  ret void
-}
-
-define void @Microsoft__Quantum__Intrinsic__H__ctl(%Array* %__controlQubits__, %Qubit* %qb) {
-entry:
-  call void @__quantum__rt__array_update_alias_count(%Array* %__controlQubits__, i64 1)
-  call void @__quantum__qis__h__ctl(%Array* %__controlQubits__, %Qubit* %qb)
-  call void @__quantum__rt__array_update_alias_count(%Array* %__controlQubits__, i64 -1)
-  ret void
-}
-
-declare void @__quantum__qis__h__ctl(%Array*, %Qubit*)
-
-define void @Microsoft__Quantum__Intrinsic__H__ctladj(%Array* %__controlQubits__, %Qubit* %qb) {
-entry:
-  call void @__quantum__rt__array_update_alias_count(%Array* %__controlQubits__, i64 1)
-  call void @__quantum__qis__h__ctl(%Array* %__controlQubits__, %Qubit* %qb)
-  call void @__quantum__rt__array_update_alias_count(%Array* %__controlQubits__, i64 -1)
-  ret void
-}
-
-define %Result* @Microsoft__Quantum__Intrinsic__Measure__body(%Array* %bases, %Array* %qubits) {
-entry:
-  call void @__quantum__rt__array_update_alias_count(%Array* %bases, i64 1)
-  call void @__quantum__rt__array_update_alias_count(%Array* %qubits, i64 1)
-  %0 = call %Result* @__quantum__qis__measure__body(%Array* %bases, %Array* %qubits)
-  call void @__quantum__rt__array_update_alias_count(%Array* %bases, i64 -1)
-  call void @__quantum__rt__array_update_alias_count(%Array* %qubits, i64 -1)
-  ret %Result* %0
-}
-
-attributes #0 = { "EntryPoint" }
